src/test/java/com/devcircus/java/algorithms/sort/quicksort/QuickTest.java
package com.devcircus.java.algorithms.sort.quicksort;

import com.devcircus.java.algorithms.sort.BaseSortTest;
import org.junit.Test;
import static org.junit.Assert.*;
import org.junit.Before;

/**
 *
 * @author Adrian Novegil <adrian.novegil@gmail.com>
 */
public class QuickTest extends BaseSortTest{

    private Quick sortter;
    
    @Before
    public void setUp() {
        sortter = new Quick();
    }
    
    @Test
    public void testSortWithIntegers() {
        assertArrayEquals(sortter.sort(getIntegerArrayToSort()), getSortedIntegerArray());
    }

    @Test
    public void testSortWithStrings() {
        assertArrayEquals(sortter.sort(getStringArrayToSort()), getSortedStringArray());
    }
}
